Add placeholders for when images are loading
Previous behavior added the images as they came, without a placeholder for
each, just a ProgressView while no image was returned. Now, there is a
ProgressView for each image and their places are defined when the placeholders
are created.

The solution achieve that with the "PhotoUpdate" enum, which allows the dev to
pass meaning information from the download site to the processing site,
indicating whether the download started, ended or the data was obtained from
local storage.